refactor:useChats로 변경
